update discount-package module
更新 優惠組合 模組的相關function

+++ conference-admin/src/main/java/tw/com/conference/pojo/entity/DiscountPackage.java
@@ -58,5 +58,5 @@ public class DiscountPackage implements Serializable {
     @Schema(description = "邏輯刪除,預設為0活耀,1為刪除")
     @TableField("is_deleted")
     @TableLogic
-    private Byte isDeleted;
+    private Integer isDeleted;
 }
